requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9

langchain==0.2.6
langchain-community==0.2.6
langchain-google-genai==1.0.10
langchain-mongodb==0.1.6

pymongo==4.7.0
chromadb==0.5.3
pypdf==4.2.0
python-dotenv==1.0.1
pydantic==2.7.4